Moving workers to a single file
