+++ whedon_api.rb
@@ -6,7 +6,6 @@ require 'fileutils'
 require 'json'
 require 'octokit'
 require 'rest-client'
-require 'sidekiq'
 require 'sinatra/config_file'
 require 'whedon'
 require 'yaml'
